+++ usvg/src/geom.rs
@@ -94,48 +94,6 @@ impl TransformFromBBox for Transform {
     }
 }
 
-/// Line representation.
-#[allow(missing_docs)]
-#[derive(Clone, Copy, Debug)]
-pub(crate) struct Line {
-    pub x1: f64,
-    pub y1: f64,
-    pub x2: f64,
-    pub y2: f64,
-}
-
-impl Line {
-    /// Creates a new line.
-    #[inline]
-    pub fn new(x1: f64, y1: f64, x2: f64, y2: f64) -> Line {
-        Line { x1, y1, x2, y2 }
-    }
-
-    /// Calculates the line length.
-    #[inline]
-    pub fn length(&self) -> f64 {
-        let x = self.x2 - self.x1;
-        let y = self.y2 - self.y1;
-        (x * x + y * y).sqrt()
-    }
-
-    /// Sets the line length.
-    pub fn set_length(&mut self, len: f64) {
-        let x = self.x2 - self.x1;
-        let y = self.y2 - self.y1;
-        let len2 = (x * x + y * y).sqrt();
-        let line = Line {
-            x1: self.x1,
-            y1: self.y1,
-            x2: self.x1 + x / len2,
-            y2: self.y1 + y / len2,
-        };
-
-        self.x2 = self.x1 + (line.x2 - line.x1) * len;
-        self.y2 = self.y1 + (line.y2 - line.y1) * len;
-    }
-}
-
 // TODO: remove
 /// A 2D point representation.
 #[derive(Clone, Copy)]