+++ CHANGELOG.md
@@ -13,6 +13,9 @@ This changelog also contains important changes in dependencies.
 - Allow writing resvg output to stdout.
 - Allow disabling text kerning using `kerning="0"` and `style="font-kerning:none"`.
 
+### Changed
+- Disable focal point correction on radial gradients to conform with SVG 2.
+
 ### Fixed
 - Do not clip nested `svg` when only the `viewBox` attribute is present.